+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Each release is divided into the following main categories:
 
 ## v2.0.x (under development)
 ### NEW
+- Added script for browser recognition/suggestions  
+import `'./modules/browser-recognition';` to app.js  
+include `'includes/browser-check.html'` to your base_root template
+
 - Added font-cropping mixin
     Example:
     ```SCSS